+++ Sources/Presenter/Coordinator/Course/AddPlanCoordinator.swift
@@ -1,5 +1,5 @@
 //
-//  EditCourseCoordinator.swift
+//  AddPlanCoordinator.swift
 //  ComeIt
 //
 //  Created by 김승창 on 2022/11/19.
@@ -10,7 +10,7 @@ import CoreLocation
 import Foundation
 import UIKit
 
-protocol EditCourseCoordinating: CourseFlowCoordinator {
+protocol AddPlanCoordinating: CourseFlowCoordinator {
     /// 장소를 선택하기 위한 지도 화면으로 이동합니다.
     /// - Parameter courseRequestDTO: 코스 관련 API 통신을 위한 DTO
     func pushToCourseMapView(_ courseRequestDTO: CourseRequestDTO)
@@ -39,7 +39,7 @@ protocol EditCourseCoordinating: CourseFlowCoordinator {
     func popViewController()
 }
 
-final class EditCourseCoordinator: CourseFlowCoordinator {
+final class AddPlanCoordinator: CourseFlowCoordinator {
     weak var navigationController: UINavigationController?
     
     init(navigationController: UINavigationController?) {
@@ -48,20 +48,19 @@ final class EditCourseCoordinator: CourseFlowCoordinator {
     
     func start(_ courseRequestDTO: CourseRequestDTO) {
         let viewModel = CourseTitleViewModel(coordinator: self, courseRequestDTO: courseRequestDTO)
-        let viewController = CourseTitleViewController(type: .editCourse, viewModel: viewModel)
+        let viewController = CourseTitleViewController(type: .addPlan, viewModel: viewModel)
         
         viewController.hidesBottomBarWhenPushed = true
-        self.navigationController?.navigationBar.isHidden = false
         
         self.navigationController?.pushViewController(viewController, animated: true)
     }
 }
 
 // MARK: - Coordinating
-extension EditCourseCoordinator: EditCourseCoordinating {
+extension AddPlanCoordinator: AddPlanCoordinating {
     func pushToCourseMapView(_ courseRequestDTO: CourseRequestDTO) {
         let viewModel = CourseMapViewModel(coordinator: self, courseRequestDTO: courseRequestDTO)
-        let viewController = CourseMapViewController(type: .editCourse, viewModel: viewModel)
+        let viewController = CourseMapViewController(type: .addPlan, viewModel: viewModel)
         
         self.navigationController?.pushViewController(viewController, animated: true)
     }
@@ -89,8 +88,8 @@ extension EditCourseCoordinator: EditCourseCoordinating {
     }
     
     func pushToCompleteView(_ courseRequestDTO: CourseRequestDTO) {
-        let viewModel = CourseCompleteViewModel(coordinator: self, courseRequestDTO: courseRequestDTO)
-        let viewController = CourseCompleteViewController(type: .editCourse, viewModel: viewModel)
+        let viewModel = PlanCompleteViewModel(coordinator: self, courseRequestDTO: courseRequestDTO)
+        let viewController = PlanCompleteViewController(viewModel: viewModel)
         
         self.navigationController?.pushViewController(viewController, animated: true)
     }
