+++ Sources/Presenter/Home/Course/Components/Annotation/PlaceSearchResultAnnotationView.swift
