+++ Sources/Presenter/Home/Course/Components/Annotation/StarAnnotationView.swift
