+++ Sources/Presenter/Home/Course/Components/Label/PaddingLabel.swift
