+++ Sources/Presenter/Home/Course/CourseMap/Cell/MapPlaceTableViewCell.swift
