+++ Sources/Presenter/Home/Course/CourseMap/View/PlaceDetailView.swift
