+++ Sources/Presenter/Home/Course/CourseMap/View/PlaceListView.swift
