+++ Sources/Presenter/Home/Course/PlaceSearch/Cell/PlaceSearchTableViewCell.swift
