+++ Sources/Presenter/Home/Course/PlaceSearch/View/EmptyResultView.swift
