+++ Sources/Presenter/Home/Course/PlaceSearchResultMap/PlaceSearchResultMapViewController.swift
