+++ Sources/Presenter/Home/Course/PlaceSearchResultMap/PlaceSearchResultMapViewModel.swift
