--- Sources/Presenter/Home/FollowerHome/FollowerHomeViewController.swift
@@ -1,31 +1,0 @@
-//
-//  FollowerHomeViewController.swift
-//  MatStar
-//
-//  Created by uiskim on 2022/10/17.
-//  Copyright © 2022 Try-ing. All rights reserved.
-//
-
-import UIKit
-
-import SnapKit
-
-class FollowerHomeViewController: UIViewController {
-    
-    private let mainLabel: UILabel = {
-        let label = UILabel()
-        label.text = "친구 별자리"
-        label.textColor = .black
-        return label
-    }()
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        view.backgroundColor = .white
-        view.addSubview(mainLabel)
-        mainLabel.snp.makeConstraints { make in
-            make.center.equalToSuperview()
-            make.size.equalTo(100)
-        }
-    }
-}
